move TestDateTimePropertyValuesComparer
